% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_or.R
\name{calc_or}
\alias{calc_or}
\title{Calculates Odds Ratio}
\usage{
calc_or(treatment, y, n, treatment_ref_lvl = NULL, y_ref_lvl = NULL)
}
\arguments{
\item{treatment}{A vector with two unique level or value.}

\item{y}{A vector with two unique level or value.}

\item{n}{count for each combination of treatment and y.}

\item{treatment_ref_lvl}{Reference Level of treatment variable.}

\item{y_ref_lvl}{Reference level of y variable.}
}
\value{
A numeric value of odds ratio
}
\description{
\code{calc_or} calculates odds ratio for Binary Outcome (y) and Binary
treatment variable, for given a vector of treatment levels, a vector of
possible outcome y and the count for each combination of treatment and y.
and \code{calc_or} supports tidy-selection.
}
