% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rr.R
\name{calc_rr}
\alias{calc_rr}
\title{Calculates Risk Ratio.}
\usage{
calc_rr(treatment, y, n, treatment_ref_lvl = NULL, y_ref_lvl = NULL)
}
\arguments{
\item{treatment}{A vector with two unique level or value.}

\item{y}{A vector with two unique level or value.}

\item{n}{count for each combination of treatment and y.}

\item{treatment_ref_lvl}{Reference Level of treatment variable.}

\item{y_ref_lvl}{Reference level of y variable.}
}
\value{
A numeric value of risk ratio
}
\description{
\code{calc_rr} calculates risk-ratio for Binary Outcome (y) and Binary
treatment variable, for given a vector of treatment levels, a vector of
possible outcome y and the count for each combination of treatment and y.
\code{calc_rr} and supports tidy-selection.
}
