% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_risk.R
\name{causal_risk}
\alias{causal_risk}
\title{Calculate measures of causal effects for conditionally randomized experiments}
\usage{
causal_risk(A, L, Y, data = NULL)
}
\arguments{
\item{A}{Vector of treatment assignment. A = 1, if the individual received the treatment; and A = 0, if the individual did not.}

\item{L}{Vector of the values of L, on which the experiment is conditionally randomized.}

\item{Y}{Vector of outcome values.  Y = 1, if the individual died; and A = 0, otherwise.}

\item{data}{(Optional) The data frame containing the A, L, and Y variables by their exact names. If this argument is provided, then the A, L, and Y arguments become optional.}
}
\value{
Measures of causal effects (e.g., risk difference, risk ratio, and odds ratio).
}
\description{
This function computes the causal risk measures in a conditionally randomized experiment via standardization.
}
\examples{
L <- rep(x = c(0,1), times = c(7, 13))
A <- c(0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1)
Y <- c(0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0)
romans <- data.frame(L = L, A = A, Y = Y)
require(dplyr)
causal_risk(A = A, L = L, Y = Y)
causal_risk(data = romans) # both are correct

}
