% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_or.R
\name{dcalc_or}
\alias{dcalc_or}
\title{Calculates Odds Ratio from a given data frame}
\usage{
dcalc_or(
  data,
  treatment,
  y,
  treatment_ref_lvl = NULL,
  y_ref_lvl = NULL,
  group = NULL
)
}
\arguments{
\item{data}{A data frame containing treatment and outcome (y) column.}

\item{treatment}{treatment column with two unique level or value.}

\item{y}{outcome column with two unique level or value.}

\item{treatment_ref_lvl}{Reference Level of treatment column.}

\item{y_ref_lvl}{Reference level of outcome (y) column.}

\item{group}{group column.}
}
\value{
A tibble containing the odds ratio
}
\description{
\code{dcalc_or} calculates the odds ratio directly from a data frame and supports
tidy-selection.
}
