% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rr.R
\name{dcalc_rr}
\alias{dcalc_rr}
\title{Calculates Risk Ratio from a given data frame}
\usage{
dcalc_rr(
  data,
  treatment,
  y,
  treatment_ref_lvl = NULL,
  y_ref_lvl = NULL,
  group = NULL
)
}
\arguments{
\item{data}{A data frame containing treatment and outcome (y) column.}

\item{treatment}{treatment column with two unique level or value.}

\item{y}{outcome column with two unique level or value.}

\item{treatment_ref_lvl}{Reference Level of treatment column.}

\item{y_ref_lvl}{Reference level of outcome (y) column.}

\item{group}{group column.}
}
\value{
A tibble containing the risk ratio.
}
\description{
\code{dcalc_rr} calculates the risk ratio directly from data and supports
tidy-selection.
}
